More comments and refactoring cleanup
(cherry picked from commit 3012e1a)
